estrutura/modularización de codigo
