+++ main.py
@@ -1,0 +1,5 @@
+# Archivo principal de arranque de la aplicación
+from gui.window import main
+
+if __name__ == "__main__":
+    main()
