+++ resources/icons/.keep
@@ -1,0 +1,2 @@
+# Carpeta para íconos de la app
+# Ejemplo: Zoo.ico
