+++ resources/images/.keep
@@ -1,0 +1,2 @@
+# Carpeta para imágenes usadas en la app
+# Ejemplo: logo, capturas, etc.
