+++ resources/images/__init__.py
@@ -1,0 +1,1 @@
+# resources/images/__init__.py
