+++ resources/templates/.keep
@@ -1,0 +1,2 @@
+# Carpeta para plantillas y archivos base .docx
+# Ejemplo: plantillaAFU2505.docx
