+++ tests/test_logic.py
@@ -1,0 +1,2 @@
+# Pruebas unitarias para la lógica de negocio
+# Aquí puedes agregar tests para app/logic.py
